pyproject.toml
[project]
name = "bookmyshow"
version = "0.1.0"
description = "A movie booking platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.9",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

# --- Ruff Configuration ---
[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
# E = pycodestyle, F = Pyflakes, I = isort (import sorting)
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
